+++ src/mac_agents_manager/constants.py
@@ -1,0 +1,3 @@
+"""Shared constants for mac_agents_manager."""
+
+MAM_LABEL = "user.productivity.mac_agents_manager"